requirements.txt
streamlit
openai
elevenlabs
SpeechRecognition
audio-recorder-streamlit
google-generativeai
